Refactoring: Shared Ranking Between Games
Changes Made

Creation of the ranking.py module to centralize the /ranking command and ranking display logic.
Removal of the /ranking command and the ranking function from the guess_who.py file.
Adaptation of the bot to load the new ranking extension, allowing shared use across different games (Guess Who, Arkdle, etc.).
Maintained the scoring and ranking display logic for all users.
Improved code organization and modularity, facilitating future integrations and maintenance.
