+++ bot.py
@@ -14,6 +14,7 @@ async def on_ready():
 # Load the slash commands extension
 bot.load_extension("commands.guess_who")
 bot.load_extension("commands.arkdle")
+bot.load_extension("commands.ranking")
 
 print("Iniciando o bot...")
 bot.start()
