+++ commands/guess_who.py
@@ -142,16 +142,6 @@ async def reveal_operator(ctx, client):
     reset_round()
 
 
-async def show_ranking(ctx):
-    scores = load_scores()
-    if not scores:
-        await ctx.send("Ninguém acertou nenhum operador ainda.")
-        return
-    ranking = sorted(scores.items(), key=lambda x: x[1]["pontos"], reverse=True)
-    msg = "**🏆 Ranking de Pontuação:**\n"
-    for i, (user_id, data) in enumerate(ranking, 1):
-        msg += f"{i}. {data['username']} — {data['pontos']} ponto(s)\n"
-    await ctx.send(msg)
 
 
 class GuessWhoGame(interactions.Extension):
@@ -188,11 +178,6 @@ class GuessWhoGame(interactions.Extension):
     async def reveal(self, ctx: interactions.SlashContext):
         await reveal_operator(ctx, self.client)
 
-    @interactions.slash_command(
-        name="ranking", description="Exibe o ranking de pontuação dos usuários."
-    )
-    async def ranking(self, ctx: interactions.SlashContext):
-        await show_ranking(ctx)
 
 
 def setup(client):
